requirements.txt
opencv-python
numpy
pillow
paddleocr
scikit-learn
deep-translator
langdetect
wordninja
libretranslatepy
streamlit
PyMuPDF